requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.2
django-enumfields==1.0.0
django-filter==2.2.0
django-rest==0.0.1
djangorestframework==3.11.0
drf-enum-field==0.9.3
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
importlib-metadata==1.4.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.1.0
openapi-codec==1.3.2
packaging==20.0
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytz==2019.3
qrcode==6.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.8
zipp==1.0.0
